# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "part of tikkichest",
        "title": "Tikkichest notifications service",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/notifications/{userID}": {
            "post": {
                "description": "turn on notifications for selected user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Notifications mode on",
                "parameters": [
                    {
                        "description": "user information: id and telegram username",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "turn off notifications for selected user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Notifications mode off",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "profile id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "change username for user with notifications mode on",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Change username",
                "parameters": [
                    {
                        "description": "user information: id and telegram username",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "telegram_username": {
                    "description": "Telegram username",
                    "type": "string"
                }
            }
        }
    }
}